add zustand and set up store
