+++ package.json
@@ -50,7 +50,8 @@
     "react-switch": "^7.0.0",
     "slick-carousel": "^1.8.1",
     "tailwindcss": "^3.2.4",
-    "typescript": "4.9.4"
+    "typescript": "4.9.4",
+    "zustand": "^4.4.4"
   },
   "devDependencies": {
     "@types/dompurify": "^3.0.1",
