--- store/services/setttings.ts
@@ -1,8 +1,0 @@
-export interface AccountSettings {
-  lastName: string;
-  firstName: string;
-  location: string;
-  phone: string;
-  email: string;
-  profileImg: object;
-}
